add fo_to_x_y & point_in_direction & change_x_by & change_y_by & set_x_to & set_y_to

+++ sources/constants.cpp
@@ -21,6 +21,6 @@ bool Limit_CharacterX (Character &sprite){
 }
 
 bool Limit_CharacterY(Character &sprite){
-    if(sprite.y <= stage.h / 2 - sprite.height / 2 && sprite.y >= -stage.h / 2 + sprite.height / 2) return true;
+    if(sprite.y <= (double)stage.h / 2 - sprite.height / 2 && sprite.y >= -(double)stage.h / 2 + sprite.height / 2) return true;
     return false;
 }
